% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadTemplate.R
\name{download_clickStories_template}
\alias{download_clickStories_template}
\title{Download the clickStories Template}
\usage{
download_clickStories_template(destfile = NULL, overwrite = FALSE)
}
\arguments{
\item{destfile}{A character string specifying the destination file path.
If NULL, the function will return the file path within the package.}

\item{overwrite}{Logical. If TRUE, overwrite an existing file at \code{destfile}.}
}
\value{
Returns the path to the file. If \code{destfile} is provided, the file is copied to the specified location.
}
\description{
This function allows users to access or download the clickStories template Excel file.
}
\examples{
# Copy the template to your working directory
download_clickStories_template("clickStories_template.xlsx")

# Get the file path within the package
download_clickStories_template()
}
