% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generatePanels.R
\name{transform_figma_embed}
\alias{transform_figma_embed}
\title{Transform Figma Embed URL to Prototype Format}
\usage{
transform_figma_embed(embed_code)
}
\arguments{
\item{embed_code}{The iframe embed code containing a Figma embed URL.}
}
\value{
A modified iframe embed code with the required parameters for the Figma prototype.
If the embed is not a Figma embed, the original embed code is returned unaltered.
}
\description{
This internal helper function transforms Figma iframe embed URLs into the
standardized prototype format. If the Figma URL is a design link, it is converted
to a prototype link, and parameters such as \code{scaling}, \code{content-scaling}, \code{hide-ui},
and \code{show-proto-sidebar} are adjusted or added as necessary.
}
\examples{
\dontrun{
figma_embed <- '<iframe style="border: 1px solid rgba(0, 0, 0, 0.1);" width="800" height="450" src="https://embed.figma.com/design/3BByo3UqianyYN2GaDmePf/Profile-Data-Stories---Brand-Colors?node-id=38-1173&embed-host=share" allowfullscreen></iframe>'
transformed_embed <- transform_figma_embed(figma_embed)
}

}
\keyword{internal}
