% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateStory.R
\name{create_story}
\alias{create_story}
\title{Create a click through data story using Quarto revealjs}
\arguments{
\item{story_title}{The title of the presentation.}

\item{logo}{The path to the logo image. If provided, it will be copied to the "images" directory.}

\item{filename}{The filename for the generated Quarto \code{.qmd} document. Default is "story.qmd".}

\item{style}{The path to a custom SCSS file for styling the presentation. If provided, it will be copied to the same directory as the \code{.qmd} file.}

\item{...}{Panel definitions. Each panel is defined as a list containing:
\item{name}{A keyword or two (hyphenated) to use as the panel's url when published.}
\item{takeaway}{The larger takeaway sentence that appears first for each panel.}
\item{text}{The additional text to provide context or explanations for the panel.}
\item{vizType}{The type of visualization to include in the second column. Possible values: "embed" (for embedding HTML content), "image-link" (for an external image), "image" (for a local image).}
\item{viz}{The content of the visualization. It can be an HTML embed code, a link to an external image, or the path to a local image.}
\item{alt}{The alt text for the image or image-link, used for accessibility and image descriptions.}}
}
\value{
Creates a Quarto \code{.qmd} file that can be used to render a Reveal.js presentation.
}
\description{
This function generates a Quarto Reveal.js presentation with custom slides, including images, text, and visualizations.
Each panel can have two columns, one for text and the other for a visualization, which could be an embedded HTML object, an external image, or a local image.
}
\examples{
create_story(
  story_title = "Data Story Title",
  logo = NULL,
  filename = "story.qmd",
  style = NULL,
  list(name = "keyword1",
    takeaway = "This is the Main Takeaway of the panel.",
    text = "This is additional text that gives the reader additional context.",
    vizType = "embed",
    viz = '<div class="flourish-embed flourish-chart" data-src="visualisation/11597006"><script src="https://public.flourish.studio/resources/embed.js"></script></div>'),
  list(name = "keyword2",
    takeaway = "New panel, new main takeaway, summary sentence, etc.",
    text = "This is more text that gives the reader additional context.",
    vizType = "image-link",
    viz = "https://www.naccho.org/uploads/body-images/public-health-101-infographic-no-logo.jpg",
    alt = "This is the alt text description for the shared image link"),
  list(name = "keyword3",
    takeaway = "New panel, new main takeaway, summary sentence, etc.",
    text = "This is more text that gives the reader additional context.",
    vizType = "image",
    viz = "images/test.png",
    alt = "The alt text for the associated image.")
)
}
